AI Copilot Dashboard/backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
openai>=1.10.0,<2.0.0
pydantic==2.4.2
langchain>=0.1.0
langchain-core>=0.1.16
langchain-community>=0.0.13
langchain-openai>=0.0.5
langchain-text-splitters>=0.0.1
langsmith>=0.0.77
chromadb==0.4.22
pypdf==3.17.4
tiktoken==0.5.2
beautifulsoup4>=4.12.0
typing-extensions>=4.5.0 